Break location bar out of header, use redux for state
This will come in handy when we start doing real location detection
